requirements.txt
# Core Web Framework
Flask==3.1.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==2.3.4
python-dateutil==2.8.2

# Statistical Analysis
scipy==1.16.2

# Sentiment Analysis (ML/AI)
transformers==4.35.2
torch==2.1.1

# Utilities
python-dotenv==1.0.0

# Production Web Server (for deployment)
gunicorn==21.2.0

